Move all classes into packages
