+++ src/main/java/duke/main/Main.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.main;
 
+import duke.ui.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -24,4 +25,5 @@ public class Main extends Application {
             e.printStackTrace();
         }
     }
+
 }
