+++ src/main/java/duke/storage/Storage.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.storage;
 
+import duke.exception.DukeException;
 import duke.task.Task;
 import duke.task.Todo;
 import duke.task.Deadline;
