+++ src/main/java/duke/task/Event.java
@@ -1,7 +1,6 @@
 package duke.task;
 
-import duke.DukeException;
-import duke.task.Task;
+import duke.exception.DukeException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
