+++ src/main/java/duke/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.ui;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
